requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.0.6
django-allauth==65.3.1
django-celery-results==2.5.1
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
kombu==5.4.2
mysql==0.0.3
mysqlclient==2.2.4
openai==1.59.6
packaging==24.1
prompt_toolkit==3.0.48
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
celery==5.4.0
django-celery-results==2.5.1
redis==5.2.1
djangorestframework_simplejwt==5.4.0
social-auth-app-django==5.4.2
PyGithub>=2.0.0